Fix build by fixing asmdefs, removing ones I can't quickly get working, and rearranging code to exclude editor-only code from builds
