+++ Assets/Photon Unity Networking/Photon.asmdef.disabled.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 12d5b6bccb9b57c4fbf287eb123407e1
-AssemblyDefinitionImporter:
+guid: 1c6852521ed38ad4da762ba7e52108b5
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
