+++ Assets/_scripts/CastIronChat.SpyVsSpy.asmdef.disabled.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: 84b315593e39c6744b5c8906f134de65
-folderAsset: yes
+guid: 518a637f3d622d944a9b2d426463c4eb
 DefaultImporter:
   externalObjects: {}
   userData:
