+++ Assets/_scripts/Utilities/Codegen/Editor/CastIronChat.CodeGen.Editor.asmdef.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a4c42bf579145634e82c8cc6f1dd3781
+guid: 98a4f6c27ed942e4bb9afb311657af04
 AssemblyDefinitionImporter:
   externalObjects: {}
   userData:
