--- Assets/_scripts/Utilities/Codegen/codegen.asmdef
@@ -1,3 +1,0 @@
-﻿{
-	"name": "codegen"
-}
