--- Assets/_scripts/Utilities/ComponentAttributes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 270344ad1adf2ec4c98473adc2122d3a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
