+++ Assets/_scripts/Utilities/Editor/CastIronChat.Utilities.Editor.asmdef
@@ -8,7 +8,9 @@
     "allowUnsafeCode": false,
     "overrideReferences": false,
     "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
+    "autoReferenced": false,
+    "defineConstraints": [
+        "UNITY_EDITOR"
+    ],
     "versionDefines": []
 }
