+++ Assets/_scripts/Utilities/Editor/DescriptionEditor.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 426b6e8bd3e245349c1e6e1b54344ad9
+timeCreated: 1594064161
